Make SS14ContentCollection generic(er)
